commit is a funny word.

+++ excluir-usuario.php
@@ -2,46 +2,53 @@
 session_start();
 require 'logica-autenticacao.php';
 
-$titulo_pagina = 'Página de exclusão de usuários';
-require 'cabecalho.php';
+if (!autenticado()) {
+    $_SESSION['restrito'] = true;
+    redireciona();
+    die();
+}
 
 require 'conexao.php';
 
 $id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_SPECIAL_CHARS);
 
-echo "<p><b>ID:</b> $id</p>";
+if ($conf["debug"]) {
+    echo "<p><b>ID:</b> $id</p>";
+}
+
+if (id_usuario() != $id) {
+    $_SESSION["result"] = false;
+    $_SESSION["erro"] = "Você está tentando excluir um usuário que não é o seu.";
+    $_SESSION["msg_erro"] = "Operação não permitida.";
+    redireciona("listagem-usuario.php");
+    die();
+}
 
 $sql = "DELETE FROM usuarios WHERE id = ?";
 
-$stmt = $conn->prepare($sql);
-$result = $stmt->execute([$id]);
+try {
+    $stmt = $conn->prepare($sql);
+    $result = $stmt->execute([$id]);
+} catch (Exception $e) {
+    $result = false;
+    $error = $e->getMessage();
+}
 
 $count = $stmt->rowCount();
 
-if ($result && $count >= 1) {
-    // deu certo o delete
-?>
-    <div class='alert alert-success' role='alert'>
-        <h4>Registro excluido com sucesso!</h4>
-    </div>
-<?php
-} elseif ($count == 0) {
-?>
-    <div class='alert alert-danger' role='alert'>
-        <h4>Falha ao efetuar exclusão.</h4>
-        <p>Não foi encontrado nenhum registro com o ID = <?= $id ?>.</p>
-    </div>
-<?php
-} else {
-    // não deu certo, erro
-    $errorArray = $stmt->errorInfo();
-?>
-    <div class='alert alert-danger' role='alert'>
-        <h4>Falha ao efetuar gravação.</h4>
-        <p><?= $errorArray[2]; ?></p>
-    </div>
-<?php
-}
-require 'rodape.php';
+$_SESSION["result"] = $result;
 
-?>
+if ($result && $count == 0) {
+    $result = false;
+    $_SESSION["msg_erro"] = "Falha ao efetuar exclusão";
+    $_SESSION["erro"] = "Não foi encontrado nenhum registro com o ID ($id).";
+} elseif ($result) {
+    //$_SESSION["msg_sucesso"] = "Registro excluido com sucesso!";
+    redireciona("sair.php");
+    exit;
+} else {
+    $_SESSION["msg_erro"] = "Falha ao efetuar exclusão.";
+    $_SESSION["erro"] = $error;
+}
+
+redireciona("listagem-usuario.php");
