+++ listagem-usuario.php
@@ -1,5 +1,4 @@
 <?php
-
 session_start();
 require 'logica-autenticacao.php';
 
@@ -36,18 +35,30 @@ $stmt = $conn->query($sql);
                 <td><?=$row["id"]?></td>
                 <td><?=$row["nome"]?></td>
                 <td><?=$row["email"]?></td>
-                    </a>
                 </td>
                 <?php
                 if (autenticado()) {
                 ?>
                 <td>
-                    <a class="btn btn-danger btn-sm" 
-                    href="excluir-usuario.php?id=<?=$row["id"]?>"
-                    onclick="if(!confirm('Tem certeza que deseja excluir?')) return false;">
+                    <?php
+                    if (id_usuario() == $row["id"]) {
+                    ?>
+                        <a class="btn btn-danger btn-sm" 
+                        href="excluir-usuario.php?id=<?=$row["id"]?>"
+                        onclick="if(!confirm('Tem certeza que deseja excluir?')) return false;">
                         <span data-feather="trash-2"></span>
                         Excluir
-                    </a>
+                        </a>
+                    <?php
+                    }else{
+                    ?>
+                        <button class="btn btn-secondary btn-sm" href="#" disabled>
+                        <span data-feather="trash-2"></span>
+                        Excluir
+                        </button>
+                    <?php
+                    }
+                    ?>
                 </td>
                 <?php
                 }
@@ -60,7 +71,34 @@ $stmt = $conn->query($sql);
     </table>
 </div>
 <?php
-
+if (isset($_SESSION["result"])) {
+    if ($_SESSION["result"]) {
+?>
+    <div class="row mt-3">
+        <div class="col-8">
+            <div class="alert alert-success mt-3" role="alert">
+                <h4><?= $_SESSION["msg_sucesso"] ?></h4>
+            </div>
+        </div>
+    </div>
+<?php
+    unset($_SESSION["msg_sucesso"]);
+    } else {
+?>
+    <div class="row mt-3">
+        <div class="col-8">
+            <div class="alert alert-danger mt-3" role="alert">
+                <h4><?= $_SESSION["msg_erro"] ?></h4>
+                <p><?= $_SESSION["erro"] ?></p>
+            </div>
+        </div>
+    </div>
+<?php
+    unset($_SESSION["msg_erro"]);
+    unset($_SESSION["erro"]);
+    }
+    unset($_SESSION["result"]);
+}
     require 'rodape.php';
 
 ?>
